Consolidated some functions from minizip.c into zip.c
