+++ zip.c
@@ -105,6 +105,9 @@
 #  endif
 #endif
 
+#define WRITEBUFFERSIZE (16384)
+#define MAXFILENAME     (256)
+
 const char zip_copyright[] = " zip 1.01 Copyright 1998-2004 Gilles Vollant - http://www.winimage.com/zLibDll";
 
 typedef struct linkedlist_datablock_internal_s
@@ -2042,5 +2045,124 @@ extern int ZEXPORT zipClose2_64(zipFile file, const char* global_comment, uLong 
 #endif
     TRYFREE(zi);
 
+    return err;
+}
+
+extern uLong ZEXPORT filetime(const char *filename, tm_zip *tmzip, uLong *dostime)
+{
+    int ret = 0;
+#ifdef _WIN32
+    FILETIME ftLocal;
+    HANDLE hFind;
+    WIN32_FIND_DATAA ff32;
+
+    hFind = FindFirstFileA(filename, &ff32);
+    if (hFind != INVALID_HANDLE_VALUE)
+    {
+        FileTimeToLocalFileTime(&(ff32.ftLastWriteTime), &ftLocal);
+        FileTimeToDosDateTime(&ftLocal,((LPWORD)dostime)+1,((LPWORD)dostime)+0);
+        FindClose(hFind);
+        ret = 1;
+    }
+#else
+#if defined unix || defined __APPLE__
+    struct stat s = {0};
+    struct tm* filedate;
+    time_t tm_t = 0;
+
+    if (strcmp(filename,"-") != 0)
+    {
+        char name[MAXFILENAME+1];
+        int len = strlen(filename);
+        if (len > MAXFILENAME)
+            len = MAXFILENAME;
+
+        strncpy(name, filename, MAXFILENAME - 1);
+        name[MAXFILENAME] = 0;
+
+        if (name[len - 1] == '/')
+            name[len - 1] = 0;
+
+        /* not all systems allow stat'ing a file with / appended */
+        if (stat(name,&s) == 0)
+        {
+            tm_t = s.st_mtime;
+            ret = 1;
+        }
+    }
+
+    filedate = localtime(&tm_t);
+
+    tmzip->tm_sec  = filedate->tm_sec;
+    tmzip->tm_min  = filedate->tm_min;
+    tmzip->tm_hour = filedate->tm_hour;
+    tmzip->tm_mday = filedate->tm_mday;
+    tmzip->tm_mon  = filedate->tm_mon ;
+    tmzip->tm_year = filedate->tm_year;
+#endif
+#endif
+    return ret;
+}
+
+extern int ZEXPORT check_file_exists(const char* filename)
+{
+    FILE* ftestexist = FOPEN_FUNC(filename, "rb");
+    if (ftestexist == NULL)
+        return 0;
+    fclose(ftestexist);
+    return 1;
+}
+
+extern int ZEXPORT is_large_file(const char* filename)
+{
+    ZPOS64_T pos = 0;
+    FILE* pFile = FOPEN_FUNC(filename, "rb");
+
+    if (pFile == NULL)
+        return 0;
+
+    FSEEKO_FUNC(pFile, 0, SEEK_END);
+    pos = FTELLO_FUNC(pFile);
+    fclose(pFile);
+
+    printf("File : %s is %lld bytes\n", filename, pos);
+
+    return (pos >= 0xffffffff);
+}
+
+/* Calculate the CRC32 of a file, because to encrypt a file, we need known the CRC32 of the file before */
+extern int ZEXPORT get_file_crc(const char* filenameinzip, void *buf, unsigned long size_buf, unsigned long* result_crc)
+{
+    FILE *fin = NULL;
+    unsigned long calculate_crc = 0;
+    unsigned long size_read = 0;
+    int err = ZIP_OK;
+
+    fin = FOPEN_FUNC(filenameinzip,"rb");
+    if (fin == NULL)
+        err = ZIP_ERRNO;
+    else
+    {
+        do
+        {
+            size_read = (int)fread(buf,1,size_buf,fin);
+
+            if ((size_read < size_buf) && (feof(fin) == 0))
+            {
+                printf("error in reading %s\n",filenameinzip);
+                err = ZIP_ERRNO;
+            }
+
+            if (size_read > 0)
+                calculate_crc = crc32(calculate_crc,buf,size_read);
+        }
+        while ((err == ZIP_OK) && (size_read > 0));
+    }
+
+    if (fin)
+        fclose(fin);
+
+    printf("file %s crc %lx\n", filenameinzip, calculate_crc);
+    *result_crc = calculate_crc;
     return err;
 }
